+++ etc/dbus-serialbattery/battery.py
@@ -106,9 +106,9 @@ class Battery(ABC):
         self.cells: List[Cell] = []
         self.control_charging = None
         self.control_voltage = None
-        self.bulk_requested = False
-        self.bulk_last_reached = 0
-        self.bulk_battery_voltage = None
+        self.soc_reset_requested = False
+        self.soc_reset_last_reached = 0
+        self.soc_reset_battery_voltage = None
         self.max_battery_voltage = None
         self.min_battery_voltage = None
         self.allow_max_voltage = True
@@ -244,34 +244,36 @@ class Battery(ABC):
             self.charge_mode = "Keep always max voltage"
 
     def prepare_voltage_management(self) -> None:
-        bulk_last_reached_days_ago = (
+        soc_reset_last_reached_days_ago = (
             0
-            if self.bulk_last_reached == 0
-            else (((int(time()) - self.bulk_last_reached) / 60 / 60 / 24))
+            if self.soc_reset_last_reached == 0
+            else (((int(time()) - self.soc_reset_last_reached) / 60 / 60 / 24))
         )
-        # set bulk_requested to True, if the days are over
+        # set soc_reset_requested to True, if the days are over
         # it gets set to False once the bulk voltage was reached once
         if (
-            utils.BULK_AFTER_DAYS is not False
-            and self.bulk_requested is False
+            utils.SOC_RESET_AFTER_DAYS is not False
+            and self.soc_reset_requested is False
             and self.allow_max_voltage
             and (
-                self.bulk_last_reached == 0
-                or utils.BULK_AFTER_DAYS < bulk_last_reached_days_ago
+                self.soc_reset_last_reached == 0
+                or utils.SOC_RESET_AFTER_DAYS < soc_reset_last_reached_days_ago
             )
         ):
             """
             logger.info(
-                f"set bulk_requested to True: first time (0) or {utils.BULK_AFTER_DAYS}"
-                + f" < {round(bulk_last_reached_days_ago, 2)}"
+                f"set soc_reset_requested to True: first time (0) or {utils.SOC_RESET_AFTER_DAYS}"
+                + f" < {round(soc_reset_last_reached_days_ago, 2)}"
             )
             """
-            self.bulk_requested = True
+            self.soc_reset_requested = True
 
-        self.bulk_battery_voltage = round(utils.BULK_CELL_VOLTAGE * self.cell_count, 2)
+        self.soc_reset_battery_voltage = round(
+            utils.SOC_RESET_VOLTAGE * self.cell_count, 2
+        )
 
-        if self.bulk_requested:
-            self.max_battery_voltage = self.bulk_battery_voltage
+        if self.soc_reset_requested:
+            self.max_battery_voltage = self.soc_reset_battery_voltage
         else:
             self.max_battery_voltage = round(
                 utils.MAX_CELL_VOLTAGE * self.cell_count, 2
@@ -302,19 +304,19 @@ class Battery(ABC):
 
                     # calculate penalty sum to prevent single cell overcharge by using current cell voltage
                     if (
-                        self.max_battery_voltage != self.bulk_battery_voltage
+                        self.max_battery_voltage != self.soc_reset_battery_voltage
                         and voltage > utils.MAX_CELL_VOLTAGE
                     ):
                         # foundHighCellVoltage: reset to False is not needed, since it is recalculated every second
                         foundHighCellVoltage = True
                         penaltySum += voltage - utils.MAX_CELL_VOLTAGE
                     elif (
-                        self.max_battery_voltage == self.bulk_battery_voltage
-                        and voltage > utils.BULK_CELL_VOLTAGE
+                        self.max_battery_voltage == self.soc_reset_battery_voltage
+                        and voltage > utils.SOC_RESET_VOLTAGE
                     ):
                         # foundHighCellVoltage: reset to False is not needed, since it is recalculated every second
                         foundHighCellVoltage = True
-                        penaltySum += voltage - utils.BULK_CELL_VOLTAGE
+                        penaltySum += voltage - utils.SOC_RESET_VOLTAGE
 
             voltageDiff = self.get_max_cell_voltage() - self.get_min_cell_voltage()
 
@@ -377,30 +379,32 @@ class Battery(ABC):
 
                 self.charge_mode = (
                     "Bulk dynamic"
-                    # if self.max_voltage_start_time is None  # remove this line after testing
-                    if self.max_battery_voltage == self.bulk_battery_voltage
+                    if self.max_voltage_start_time is None
                     else "Absorption dynamic"
                 )
+
+                if self.max_battery_voltage == self.soc_reset_battery_voltage:
+                    self.charge_mode += " & SoC Reset"
 
             elif self.allow_max_voltage:
                 self.control_voltage = round(self.max_battery_voltage, 3)
                 self.charge_mode = (
-                    "Bulk"
-                    # if self.max_voltage_start_time is None  # remove this line after testing
-                    if self.max_battery_voltage == self.bulk_battery_voltage
-                    else "Absorption"
+                    "Bulk" if self.max_voltage_start_time is None else "Absorption"
                 )
+
+                if self.max_battery_voltage == self.soc_reset_battery_voltage:
+                    self.charge_mode += " & SoC Reset"
 
             else:
                 floatVoltage = round((utils.FLOAT_CELL_VOLTAGE * self.cell_count), 3)
                 chargeMode = "Float"
                 # reset bulk when going into float
-                if self.bulk_requested:
-                    # logger.info("set bulk_requested to False")
-                    self.bulk_requested = False
-                    # IDEA: Save "bulk_last_reached" in the dbus path com.victronenergy.settings
+                if self.soc_reset_requested:
+                    # logger.info("set soc_reset_requested to False")
+                    self.soc_reset_requested = False
+                    # IDEA: Save "soc_reset_last_reached" in the dbus path com.victronenergy.settings
                     # to make it restart persistent
-                    self.bulk_last_reached = current_time
+                    self.soc_reset_last_reached = current_time
                 if self.control_voltage:
                     # check if battery changed from bulk/absoprtion to float
                     if not self.charge_mode.startswith("Float"):
@@ -465,16 +469,16 @@ class Battery(ABC):
             self.charge_mode_debug += (
                 f"\nlinear_cvl_last_set: {self.linear_cvl_last_set}"
             )
-            bulk_days_ago = round(
-                (current_time - self.bulk_last_reached) / 60 / 60 / 24, 2
+            soc_reset_days_ago = round(
+                (current_time - self.soc_reset_last_reached) / 60 / 60 / 24, 2
             )
-            bulk_in_days = round(utils.BULK_AFTER_DAYS - bulk_days_ago, 2)
-            self.charge_mode_debug += "\nbulk_last_reached: " + str(
+            soc_reset_in_days = round(utils.SOC_RESET_AFTER_DAYS - soc_reset_days_ago, 2)
+            self.charge_mode_debug += "\nsoc_reset_last_reached: " + str(
                 "Never"
-                if self.bulk_last_reached == 0
-                else str(bulk_days_ago)
+                if self.soc_reset_last_reached == 0
+                else str(soc_reset_days_ago)
                 + " days ago - next in "
-                + str(bulk_in_days)
+                + str(soc_reset_in_days)
                 + "days"
             )
             # """
@@ -545,6 +549,9 @@ class Battery(ABC):
                     "Bulk" if self.max_voltage_start_time is None else "Absorption"
                 )
 
+                if self.max_battery_voltage == self.soc_reset_battery_voltage:
+                    self.charge_mode += " & SoC Reset"
+
             else:
                 # check if battery changed from bulk/absoprtion to float
                 if not self.charge_mode.startswith("Float"):
@@ -553,10 +560,10 @@ class Battery(ABC):
                 self.control_voltage = utils.FLOAT_CELL_VOLTAGE * self.cell_count
                 self.charge_mode = "Float"
                 # reset bulk when going into float
-                if self.bulk_requested:
-                    # logger.info("set bulk_requested to False")
-                    self.bulk_requested = False
-                    self.bulk_last_reached = current_time
+                if self.soc_reset_requested:
+                    # logger.info("set soc_reset_requested to False")
+                    self.soc_reset_requested = False
+                    self.soc_reset_last_reached = current_time
 
             self.charge_mode += " (Step Mode)"