+++ etc/dbus-serialbattery/dbushelper.py
@@ -522,8 +522,8 @@ class DbusHelper:
         self._dbusservice["/Alarms/HighVoltage"] = (
             self.battery.protection.voltage_high
             if (
-                self.battery.bulk_requested is False
-                and self.battery.bulk_last_reached < int(time()) - (60 * 30)
+                self.battery.soc_reset_requested is False
+                and self.battery.soc_reset_last_reached < int(time()) - (60 * 30)
             )
             else 0
         )